pyproject.toml
[tool.poetry]
name = "iq-eq-tests"
version = "0.1.0"
description = ""
authors = ["Alexey Kuzenkov <kuzenkovag@yandex.ru>"]
readme = "README.md"
packages = [{include = "iq_eq_tests"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
shortuuid = "^1.0.11"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

